CF-1294 Add a 'properties' argument to sb.launch() that handles all property overrides.
The existing 'execparams' argument has been deprecated, and the usage of 'configure' to provide overrides has also been deprecated; however, 'configure' can be used with a boolean value to disable the initial call to configure(). The legacy usage of None to disable configure() is still supported (None evaluates to false).