CF-1580 Avoid potential race conditions in domain awareness test components
There were occasional CI failures in the C++ "unaware" test, where the component ID was slightly different than the expected value. I suspect this was due to a race condition in modifying the member variables without a mutex. Since the application, DomainManager and NIC are known at initialization time, the C++, Java and Python components were updated to set the values once in constructor(). This also allows removing the time.sleep() calls in the test.