CF-1646 Handle end-of-stream correctly for disabled streams
In order to avoid a race condition between the thread re-enabling and reading from the stream, and the thread servicing the pushPacket call, additional mutual exclusion needed to be applied. There does not appear to be a measurable peformance penalty for the added locking. Along with the fix, the previously unimplemented functionality of purging the queue on disable() is now implemented.