CF-1671 Don't include Java test components in the build if Java is not enabled.
Prior to integrating the test components into the main Autotools build, each component was built individually from a script, so even if Java components failed, the rest built. Since the components are now part of a Makefile, the Java components need to be excluded to ensure that the C++ components still build.