CF-1710 Use a string sequence property in ComponentHost to control the set of preloaded libraries.
This limits the set of libraries that we try to load, but in a dynamic way (default list is in ComponentHost's PRF, but could be overridden by the Domain/Sandbox). The preloading is also more fault-tolerant, so that a bad library won't prevent ComponentHost from working. Because the mechanism is different than the normal ModuleBundle mode of operation, the "default" bundle was eliminated, as there is no concern about trying to unload the preloaded libraries.