CF-1921 Fall back to doing in-band transfers via the FIFO when shared memory is exhausted
Peformance is significantly degraded in this mode, as there is copy-in/copy-out via the FIFO for each transfer; however, the system continues to operate. Some bugs with large reads and writes in the FIFO pipe class were fixed as well. The likelihood of these bugs manifesting in the normal shared memory case was effectively zero, but they were exposed by using the FIFO for full data transfer.