Segfault on exit fix
The destructor code actually does nothing since the map gets cleared by terminate's erase of everything in the map.  Never the less, delete of a Device_impl pointer is bad, rather you should _remove_ref().  However since that code is never reached, we can add the call to _remove_ref() after releaseObject() and then set the pointer to 0 to prevent its use in the future (which TBH is never since we're in the terminate function).